backend/requirements.txt
google-genai
python-dotenv
fastapi
